requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
defusedxml==0.6.0
diff-match-patch==20181111
Django==3.0.3
django-admin==2.0.1
django-admin-rangefilter==0.6.0
django-db-views==0.0.6
django-excel-base==1.0.4
django-excel-response2==3.0.2
django-filter==2.2.0
django-import-export==2.1.0
django-six==1.0.4
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
et-xmlfile==1.0.1
jdcal==1.4.1
Markdown==3.2.2
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.3
PyJWT==1.7.1
pytz==2020.1
PyYAML==5.3.1
screen==1.0.1
six==1.15.0
sqlparse==0.3.1
tablib==2.0.0
wincertstore==0.2
xlrd==1.2.0
xlwt==1.3.0
